feat: allow controlled usage of SingleCollapse
BREAKING CHANGE: rename and change SingleCollapse props
